.NET6

+++ Lib/cs/Husty/_/MultidimensionalArray.cs
@@ -10,7 +10,7 @@ namespace Husty
 
         private readonly int[] _vectorSizes;
         private readonly int[] _dimSizes;
-        private double[] _values;
+        private float[] _values;
 
 
         // ------ constructor ------ //
@@ -20,7 +20,7 @@ namespace Husty
             if (vectorSizes.Any(x => x <= 0))
                 throw new ArgumentException("all vector sizes must be positive value.");
             _vectorSizes = vectorSizes;
-            _values = new double[_vectorSizes.Aggregate(1, (s, v) => s *= v)];
+            _values = new float[_vectorSizes.Aggregate(1, (s, v) => s *= v)];
             _dimSizes = new int[_vectorSizes.Length];
             for (int i = _vectorSizes.Length - 1; i > -1; i--)
                 _dimSizes[i] = _vectorSizes.Take(i).Aggregate(1, (s, v) => s *= v);
@@ -39,7 +39,7 @@ namespace Husty
             return _values.Length;
         }
 
-        public void SetAt(int[] location, double value)
+        public void SetAt(int[] location, float value)
         {
             if (location.Length != _vectorSizes.Length)
                 throw new ArgumentException(nameof(location));
@@ -47,14 +47,14 @@ namespace Husty
             _values[loc1D] = value;
         }
 
-        public void SetAll(double[] values)
+        public void SetAll(float[] values)
         {
             if (_values.Length != values.Length)
                 throw new ArgumentException(nameof(values));
             _values = values;
         }
 
-        public double GetAt(int[] location)
+        public float GetAt(int[] location)
         {
             if (location.Length != _vectorSizes.Length)
                 throw new ArgumentException(nameof(location));
@@ -62,30 +62,30 @@ namespace Husty
             return _values[loc1D];
         }
 
-        public double[] GetAll()
+        public float[] GetAll()
         {
             return _values;
         }
 
-        public void PlusAt(int[] location, double value)
+        public void PlusAt(int[] location, float value)
         {
             var loc1D = FlattenLocation(location);
             _values[loc1D] += value;
         }
 
-        public void MinusAt(int[] location, double value)
+        public void MinusAt(int[] location, float value)
         {
             var loc1D = FlattenLocation(location);
             _values[loc1D] -= value;
         }
 
-        public void MulAt(int[] location, double value)
+        public void MulAt(int[] location, float value)
         {
             var loc1D = FlattenLocation(location);
             _values[loc1D] *= value;
         }
 
-        public void DivAt(int[] location, double value)
+        public void DivAt(int[] location, float value)
         {
             var loc1D = FlattenLocation(location);
             _values[loc1D] /= value;
